requirements.txt
langchain==0.1.12
langchain-groq==0.0.1
python-dotenv==1.0.1
pydantic==2.6.3
rich==13.7.0 